requirements.txt
nltk
openai
pytube
fastapi
uvicorn
tenacity
tiktoken
langchain
streamlit
transformers
python-dotenv
google-auth-httplib2 
google-auth-oauthlib
langchain-google-genai
google-api-python-client
pydantic
langchain_text_splitters
beautifulsoup4
requests
python-multipart
fitz
gTTS
protobuf
SpeechRecognition
google-generativeai
frontend
PyPdf2